requirements.txt
click
numpy
pandas
Pillow